+++ Changelog.md
@@ -10,6 +10,7 @@ Compiler Features:
  * Peephole Optimizer: Remove double ``iszero`` before ``jumpi``.
  * SMTChecker: Support enums without typecast.
  * SMTChecker: Support one-dimensional arrays.
+ * SMTChecker: Support modifiers.
  * Yul Optimizer: Add rule to remove empty default switch cases
  * Yul Optimizer: Add rule to remove empty cases if no default exists
  * Yul Optimizer: Add rule to replace a switch with no cases with pop(expression)